+++ src/main/java/com/mraof/minestuck/client/gui/playerStats/EcheladderScreen.java
@@ -4,13 +4,13 @@ import com.google.common.collect.ImmutableList;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mraof.minestuck.MinestuckConfig;
 import com.mraof.minestuck.client.ClientRungData;
+import com.mraof.minestuck.item.MSItems;
 import com.mraof.minestuck.player.ClientPlayerData;
 import com.mraof.minestuck.player.Rung;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractWidget;
 import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.client.gui.narration.NarrationElementOutput;
-import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.MobEffectTextureManager;
 import net.minecraft.client.resources.language.I18n;
 import net.minecraft.network.chat.Component;
@@ -31,6 +31,7 @@ public class EcheladderScreen extends PlayerStatsScreen
 	public static final String ATTACK = "minestuck.echeladder.attack";
 	public static final String HEALTH = "minestuck.echeladder.health";
 	public static final String CACHE = "minestuck.echeladder.cache";
+	public static final String CAPTCHA = "minestuck.echeladder.captcha";
 	public static final String DAMAGE_UNDERLING = "minestuck.echeladder.damage_underling";
 	public static final String DAMAGE_UNDERLING_INCREASE = "minestuck.echeladder.damage_underling.increase";
 	public static final String PROTECTION_UNDERLING = "minestuck.echeladder.protection_underling";
@@ -179,43 +180,53 @@ public class EcheladderScreen extends PlayerStatsScreen
 	@Nullable
 	private List<Component> drawEffectIconsAndText(GuiGraphics guiGraphics, int currentRung, int mouseX, int mouseY)
 	{
-		MobEffectTextureManager sprites = this.minecraft.getMobEffectTextures();
-		TextureAtlasSprite strengthSprite = sprites.get(MobEffects.DAMAGE_BOOST);
-		TextureAtlasSprite healthSprite = sprites.get(MobEffects.HEALTH_BOOST);
+		MobEffectTextureManager effectSprites = this.minecraft.getMobEffectTextures();
 		
 		int textOffset = xOffset + 24;
 		
+		int boondollarY = 12;
+		int attackY = 30;
+		int healthY = 65;
+		int cacheY = 100;
+		int captchaY = 135;
+		
 		RenderSystem.setShaderColor(1, 1, 1, 1);
-		guiGraphics.blit(xOffset + 5, yOffset + 30, 0, 18, 18, strengthSprite);
-		guiGraphics.blit(xOffset + 5, yOffset + 84, 0, 18, 18, healthSprite);
-		guiGraphics.blit(PlayerStatsScreen.icons, xOffset + 6, yOffset + 139, 48, 64, 16, 16);
-		guiGraphics.blit(PlayerStatsScreen.icons, xOffset + 5, yOffset + 7, 238, 16, 18, 18);
+		guiGraphics.blit(PlayerStatsScreen.icons, xOffset + 5, yOffset + boondollarY - 5, 238, 16, 18, 18);
+		guiGraphics.blit(xOffset + 5, yOffset + attackY, 0, 18, 18, effectSprites.get(MobEffects.DAMAGE_BOOST));
+		guiGraphics.blit(xOffset + 5, yOffset + healthY, 0, 18, 18, effectSprites.get(MobEffects.HEALTH_BOOST));
+		guiGraphics.blit(PlayerStatsScreen.icons, xOffset + 6, yOffset + cacheY + 1, 48, 64, 16, 16);
+		guiGraphics.renderItem(MSItems.CAPTCHA_CARD.toStack(), xOffset + 5, yOffset + captchaY);
+		
 		
 		String msg = title.getString();
 		guiGraphics.drawString(font, msg, xOffset + 168 - mc.font.width(msg) / 2F, yOffset + 12, GREY, false);
 		
 		Rung.DisplayData rungData = ClientRungData.getData(currentRung);
 		
-		int attack = (int) Math.round(100 * (1 + rungData.attributes().attackBonus()));
-		guiGraphics.drawString(font, I18n.get(ATTACK), textOffset, yOffset + 30, GREY, false);
-		String attackValueText = attack + "%";
-		guiGraphics.drawString(font, attackValueText, textOffset + 2, yOffset + 39, BLUE, false);
-		
-		double health = rungData.attributes().healthBoost() / 2D;
-		guiGraphics.drawString(font, I18n.get(HEALTH), textOffset, yOffset + 84, GREY, false);
-		String healthValueText = "+" + String.format(Locale.ROOT, "%.1f", health);
-		guiGraphics.drawString(font, healthValueText, textOffset + 2, yOffset + 93, BLUE, false);
-		
-		guiGraphics.drawString(font, "=", textOffset + 1, yOffset + 12, GREY, false);    //Should this be black, or the same blue as the numbers?
-		guiGraphics.drawString(font, String.valueOf(ClientPlayerData.getBoondollars()), textOffset + 3 + mc.font.width("="), yOffset + 12, BLUE, false);
+		guiGraphics.drawString(font, "=", textOffset + 1, yOffset + boondollarY, GREY, false);    //Should this be black, or the same blue as the numbers?
+		guiGraphics.drawString(font, String.valueOf(ClientPlayerData.getBoondollars()), textOffset + 3 + mc.font.width("="), yOffset + boondollarY, BLUE, false);
 		//guiGraphics.drawString("Rep: " + ClientPlayerData.getConsortReputation(), xOffset + 75 + mc.fontRenderer.getCharWidth('='), yOffset + 12, BLUE);
 		
-		guiGraphics.drawString(font, I18n.get(CACHE), textOffset, yOffset + 138, GREY, false);
-		guiGraphics.drawString(font, String.valueOf(rungData.gristCapacity()), textOffset + 2, yOffset + 147, BLUE, false);
+		int attack = (int) Math.round(100 * (1 + rungData.attributes().attackBonus()));
+		guiGraphics.drawString(font, I18n.get(ATTACK), textOffset, yOffset + attackY, GREY, false);
+		String attackValueText = attack + "%";
+		guiGraphics.drawString(font, attackValueText, textOffset + 2, yOffset + attackY + 9, BLUE, false);
 		
-		if(mouseInBounds(mouseY, yOffset + 39, mouseX, textOffset + 2, mc.font.width(attackValueText)))
+		double health = rungData.attributes().healthBoost() / 2D;
+		guiGraphics.drawString(font, I18n.get(HEALTH), textOffset, yOffset + healthY, GREY, false);
+		String healthValueText = "+" + String.format(Locale.ROOT, "%.1f", health);
+		guiGraphics.drawString(font, healthValueText, textOffset + 2, yOffset + healthY + 9, BLUE, false);
+		
+		guiGraphics.drawString(font, I18n.get(CACHE), textOffset, yOffset + cacheY, GREY, false);
+		guiGraphics.drawString(font, String.valueOf(rungData.gristCapacity()), textOffset + 2, yOffset + cacheY + 9, BLUE, false);
+		
+		guiGraphics.drawString(font, I18n.get(CAPTCHA), textOffset, yOffset + captchaY, GREY, false);
+		String captchaValueText = String.format(Locale.ROOT, "%d", (int) rungData.attributes().captchalogueCapacity());
+		guiGraphics.drawString(font, captchaValueText, textOffset + 2, yOffset + captchaY + 9, BLUE, false);
+		
+		if(mouseInBounds(mouseY, yOffset + attackY + 9, mouseX, textOffset + 2, mc.font.width(attackValueText)))
 			return ImmutableList.of(Component.translatable(DAMAGE_UNDERLING), Component.literal(Math.round(attack * rungData.attributes().underlingDamageMod()) + "%"));
-		if(mouseInBounds(mouseY, yOffset + 93, mouseX, textOffset + 2, mc.font.width(healthValueText)))
+		if(mouseInBounds(mouseY, yOffset + healthY + 9, mouseX, textOffset + 2, mc.font.width(healthValueText)))
 			return ImmutableList.of(Component.translatable(PROTECTION_UNDERLING), Component.literal(String.format(Locale.ROOT, "%.1f", 100 * rungData.attributes().underlingProtectionMod()) + "%"));
 		return null;
 	}