+++ src/main/java/com/mraof/minestuck/inventory/captchalogue/CaptchaDeckHandler.java
@@ -4,6 +4,7 @@ import com.mraof.minestuck.Minestuck;
 import com.mraof.minestuck.MinestuckConfig;
 import com.mraof.minestuck.advancements.MSCriteriaTriggers;
 import com.mraof.minestuck.computer.editmode.ServerEditHandler;
+import com.mraof.minestuck.entity.MSAttributes;
 import com.mraof.minestuck.item.BoondollarsItem;
 import com.mraof.minestuck.item.CaptchaCardItem;
 import com.mraof.minestuck.item.MSItems;
@@ -19,7 +20,6 @@ import net.minecraft.core.NonNullList;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
@@ -28,7 +28,6 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentEffectComponents;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.level.GameRules;
-import net.minecraft.world.phys.Vec3;
 import net.neoforged.bus.api.EventPriority;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.LogicalSide;
@@ -42,6 +41,7 @@ import org.apache.logging.log4j.Logger;
 import javax.annotation.Nullable;
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.atomic.AtomicLong;
 
 //todo this class could use some spring cleaning
 @EventBusSubscriber(modid = Minestuck.MOD_ID, bus = EventBusSubscriber.Bus.GAME)
@@ -156,7 +156,7 @@ public final class CaptchaDeckHandler
 			ModusHolder modusHolder = getHolder(player);
 			newModus.initModus(modusItem, player, null, modusHolder.givenModus
 					? 0
-					: MinestuckConfig.SERVER.initialModusSize.get());
+					: getInitialModusSize(player));
 		} else
 		{
 			ModusType<?> oldType = oldModus.getType();
@@ -204,10 +204,20 @@ public final class CaptchaDeckHandler
 	
 	public static void captchalogueItem(ServerPlayer player)
 	{
-		if(canPlayerUseModus(player) && hasModus(player))
+		ItemStack stack = player.getMainHandItem();
+		if(canPlayerUseModus(player) && hasModus(player) && meetsComponentSizeLimit(stack))
 		{
-			captchalogueItem(player, player.getMainHandItem());
+			captchalogueItem(player, stack);
 		}
+	}
+	
+	public static boolean meetsComponentSizeLimit(ItemStack stack)
+	{
+		AtomicLong componentDataSize = new AtomicLong(0);
+		
+		stack.getComponents().forEach(component -> componentDataSize.addAndGet(component.value().toString().chars().sum()));
+		
+		return MinestuckConfig.SERVER.captchaComponentSize.get() >= (int) componentDataSize.get();
 	}
 	
 	public static void captchalogueItemInSlot(ServerPlayer player, int slotIndex, int windowId)
@@ -235,7 +245,7 @@ public final class CaptchaDeckHandler
 	public static void captchalogueItemCarried(ServerPlayer player)
 	{
 		AbstractContainerMenu containerMenu = player.containerMenu;
-		if(canPlayerUseModus(player) && hasModus(player) && containerMenu != null)
+		if(canPlayerUseModus(player) && hasModus(player))
 		{
 			captchalogueItem(player, containerMenu.getCarried());
 		}
@@ -320,13 +330,10 @@ public final class CaptchaDeckHandler
 		if(!stack.isEmpty())
 		{
 			AbstractContainerMenu containerMenu = player.containerMenu;
-			ItemStack otherStack = containerMenu != null ? containerMenu.getCarried() : player.getMainHandItem();
+			ItemStack otherStack = containerMenu.getCarried();
 			if(otherStack.isEmpty())
 			{
-				if(containerMenu != null)
-					containerMenu.setCarried(stack);
-				else
-					player.setItemInHand(InteractionHand.MAIN_HAND, stack);
+				containerMenu.setCarried(stack);
 			} else if(canMergeItemStacks(stack, otherStack))
 			{
 				otherStack.grow(stack.getCount());
@@ -368,7 +375,7 @@ public final class CaptchaDeckHandler
 		int cardsToKeep = switch(MinestuckConfig.SERVER.sylladexDropMode.get())
 		{
 			case ITEMS -> size;
-			case CARDS_AND_ITEMS -> MinestuckConfig.SERVER.initialModusSize.get();
+			case CARDS_AND_ITEMS -> getInitialModusSize(player);
 			case ALL -> 0;
 		};
 		
@@ -446,9 +453,7 @@ public final class CaptchaDeckHandler
 	public static Modus getModus(ServerPlayer player)
 	{
 		Optional<PlayerData> playerData = PlayerData.get(player);
-		if(playerData.isEmpty())
-			return null;
-		return playerData.get().getData(MSAttachments.MODUS_HOLDER).modus;
+		return playerData.map(data -> data.getData(MSAttachments.MODUS_HOLDER).modus).orElse(null);
 	}
 	
 	private static boolean canMergeItemStacks(ItemStack stack1, ItemStack stack2)
@@ -494,7 +499,7 @@ public final class CaptchaDeckHandler
 			return;
 		}
 		
-		modus.initModus(null, player, null, MinestuckConfig.SERVER.initialModusSize.get());
+		modus.initModus(null, player, null, getInitialModusSize(player));
 		setModus(modusHolder, player, modus);
 	}
 	
@@ -502,6 +507,11 @@ public final class CaptchaDeckHandler
 	{
 		PlayerData data = PlayerData.get(player).orElseThrow();
 		return data.getData(MSAttachments.MODUS_HOLDER);
+	}
+	
+	private static int getInitialModusSize(ServerPlayer player)
+	{
+		return (int) Math.min(MinestuckConfig.SERVER.initialModusSize.get(), player.getAttributeValue(MSAttributes.CAPTCHALOGUE_CAPACITY));
 	}
 	
 	public static class ModusHolder implements INBTSerializable<CompoundTag>