+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Added 3 bunker structures to Lands
 - Increased block diversity in structures
 - Array modus
+- Added a limit to captchaloguing cards with large amounts of stored data
 
 ### Changed
 
@@ -32,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Underlings no longer spawn naturally in consort villages and consorts no longer spawn in dungeons
 - Inventory item discharge is now divided into separate eject and launch mechanics
 - HashmapModus now can handle with localized names, tags, anvil names
+- The max amount of cards held in the modus is now handled by an Attribute, which is tied to the Echeladder
 
 ### Fixed